# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Master Class education platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  teacher: 'teacher',
  student: 'student',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  invited: 'invited',
  suspended: 'suspended',
} as const;

export type ClassStatus = typeof ClassStatus[keyof typeof ClassStatus];


export const ClassStatus = {
  active: 'active',
  draft: 'draft',
  archived: 'archived',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  avatar?: string | null;
  classesCount: number;
  createdAt: string;
}

export interface CurrentUser {
  id: number;
  clerkUserId: string;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  role: UserRole;
  status?: UserStatus;
}

export interface UserUpdate {
  /** @minLength 2 */
  name?: string;
  email?: string;
  role?: UserRole;
  status?: UserStatus;
}

export interface UserPasswordUpdate {
  /** @minLength 8 */
  password: string;
  signOutOfOtherSessions?: boolean;
}

export interface PasswordUpdateResult {
  success: boolean;
}

/**
 * @nullable
 */
export type ClassTeacher = {
  id?: number;
  name?: string;
} | null;

export interface Class {
  id: number;
  name: string;
  level: string;
  academicYear: string;
  status: ClassStatus;
  /** @nullable */
  teacher?: ClassTeacher;
  studentCount: number;
  moduleCount: number;
  createdAt: string;
}

export interface ClassInput {
  /** @minLength 2 */
  name: string;
  level: string;
  academicYear: string;
  status?: ClassStatus;
}

export interface ClassUpdate {
  /** @minLength 2 */
  name?: string;
  level?: string;
  academicYear?: string;
  status?: ClassStatus;
}

export interface Module {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  color: string;
  activityCount: number;
  editorUrl: string;
  repositoryUrl: string;
}

export interface TeacherClass {
  id: number;
  name: string;
  level: string;
  academicYear: string;
  status: ClassStatus;
  modules: Module[];
  activityCount: number;
  createdAt: string;
}

export type EditorMode = typeof EditorMode[keyof typeof EditorMode];


export const EditorMode = {
  block: 'block',
  code: 'code',
} as const;

export interface TeacherActivity {
  id: number;
  classId: number;
  moduleId: number;
  createdByUserId: number;
  title: string;
  instructions: string;
  editorMode: EditorMode;
  blockXml: string;
  sourceCode: string;
  createdAt: string;
  updatedAt: string;
}

export interface TeacherActivityInput {
  moduleId: number;
  /** @minLength 1 */
  title: string;
  instructions: string;
  editorMode: EditorMode;
  blockXml?: string;
  sourceCode?: string;
}

export interface TeacherActivityUpdate {
  moduleId?: number;
  /** @minLength 1 */
  title?: string;
  instructions?: string;
  editorMode?: EditorMode;
  blockXml?: string;
  sourceCode?: string;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  draft: 'draft',
  submitted: 'submitted',
} as const;

export interface StudentSubmission {
  id: number;
  activityId: number;
  studentId: number;
  editorMode: EditorMode;
  blockXml: string;
  sourceCode: string;
  status: SubmissionStatus;
  /** @nullable */
  submittedAt?: string | null;
  updatedAt: string;
}

export interface StudentActivity {
  id: number;
  classId: number;
  moduleId: number;
  title: string;
  instructions: string;
  editorMode: EditorMode;
  blockXml: string;
  sourceCode: string;
  module: Module;
  submission: StudentSubmission | null;
  createdAt: string;
  updatedAt: string;
}

export interface StudentClass {
  id: number;
  name: string;
  level: string;
  academicYear: string;
  status: ClassStatus;
  modules: Module[];
  activityCount: number;
  submittedCount: number;
  createdAt: string;
}

export interface StudentSubmissionInput {
  editorMode: EditorMode;
  blockXml?: string;
  sourceCode?: string;
  status: SubmissionStatus;
}

export interface TeacherActivitySubmission {
  id: number;
  activityId: number;
  studentId: number;
  studentName: string;
  studentEmail: string;
  editorMode: EditorMode;
  blockXml: string;
  sourceCode: string;
  status: SubmissionStatus;
  /** @nullable */
  submittedAt?: string | null;
  updatedAt: string;
}

export interface AssignedUser {
  id: number;
  name: string;
  email: string;
}

export interface ClassAssignments {
  classId: number;
  /** @nullable */
  teacher: AssignedUser | null;
  students: AssignedUser[];
  modules: Module[];
}

export interface TeacherAssignmentInput {
  teacherId: number;
}

export interface ModulesAssignmentInput {
  moduleIds: number[];
}

export interface StudentsAssignmentInput {
  studentIds: number[];
}

export interface Activity {
  id: number;
  title: string;
  type: string;
  time: string;
  status: string;
}

export interface DashboardSummary {
  totalUsers: number;
  totalStudents: number;
  totalTeachers: number;
  totalClasses: number;
  totalModules: number;
  activeClasses: number;
  recentActivity: Activity[];
}

export type SearchParameter = string;

export type ListUsersParams = {
search?: SearchParameter;
role?: UserRole;
status?: UserStatus;
};

export type ListClassesParams = {
search?: SearchParameter;
status?: ClassStatus;
};