# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the Master Class education platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/classes/{classId}/activities:
    get:
      operationId: listStudentClassActivities
      tags: [student]
      summary: List activities available to a student in a class
      parameters:
        - $ref: "#/components/parameters/ClassId"
      responses:
        "200":
          description: Activities and the current student's submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentActivity"
  /student/activities/{activityId}:
    get:
      operationId: getStudentActivity
      tags: [student]
      summary: Get an activity and the current student's work
      parameters:
        - $ref: "#/components/parameters/ActivityId"
      responses:
        "200":
          description: Activity and the current student's submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentActivity"
  /student/activities/{activityId}/submission:
    put:
      operationId: saveStudentSubmission
      tags: [student]
      summary: Save or submit the current student's work
      parameters:
        - $ref: "#/components/parameters/ActivityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentSubmissionInput"
      responses:
        "200":
          description: Saved student submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentSubmission"
...
components:
  parameters:
...
    ClassId:
      name: classId
      in: path
      required: true
      schema:
        type: integer
    ActivityId:
      name: activityId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Module:
      type: object
      required: [id, slug, name, shortDescription, color, activityCount, editorUrl, repositoryUrl]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        color:
          type: string
        activityCount:
          type: integer
        editorUrl:
          type: string
          format: uri
        repositoryUrl:
          type: string
          format: uri
...
    EditorMode:
      type: string
      enum: [block, code]
...
    SubmissionStatus:
      type: string
      enum: [draft, submitted]
    StudentSubmission:
      type: object
      required: [id, activityId, studentId, editorMode, blockXml, sourceCode, status, updatedAt]
      properties:
        id:
          type: integer
        activityId:
          type: integer
        studentId:
          type: integer
        editorMode:
          $ref: "#/components/schemas/EditorMode"
        blockXml:
          type: string
        sourceCode:
          type: string
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        submittedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StudentActivity:
      type: object
      required: [id, classId, moduleId, title, instructions, editorMode, blockXml, sourceCode, module, submission, createdAt, updatedAt]
      properties:
        id:
          type: integer
        classId:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        instructions:
          type: string
        editorMode:
          $ref: "#/components/schemas/EditorMode"
        blockXml:
          type: string
        sourceCode:
          type: string
        module:
          $ref: "#/components/schemas/Module"
        submission:
          oneOf:
            - $ref: "#/components/schemas/StudentSubmission"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    StudentSubmissionInput:
      type: object
      required: [editorMode, status]
      properties:
        editorMode:
          $ref: "#/components/schemas/EditorMode"
        blockXml:
          type: string
        sourceCode:
          type: string
        status:
          $ref: "#/components/schemas/SubmissionStatus"
...