# lib/api-zod/src/generated/types/studentSubmissionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Master Class education platform
 * OpenAPI spec version: 0.1.0
 */
import type { EditorMode } from './editorMode';
import type { SubmissionStatus } from './submissionStatus';

export interface StudentSubmissionInput {
  editorMode: EditorMode;
  blockXml?: string;
  sourceCode?: string;
  status: SubmissionStatus;
}